server/requirements.txt
django
djangorestframework
requests
psycopg2-binary
dj-database-url
celery
kombu
requests
black
isort
boto3
watchdog
pyyaml
argh
environs
netifaces
autoflake
